[Lua] Move HashIndexTable and Mandelbrot into separate file
Makes managing large code pieces simpler.
Keeping them as strings isn’t ideal.
