+++ benchmarks/Lua/json.lua
@@ -165,69 +165,15 @@ end
 
 end -- class JsonString
 
+local HashIndexTable
+if _VERSION < 'Lua 5.3' then
+    HashIndexTable = require'hashindextable'
+else
+    HashIndexTable = require'hashindextable.53'
+end
+
 local JsonObject = {_CLASS = 'JsonObject'} do
 setmetatable(JsonObject, {__index = JsonValue})
-
-local HashIndexTable = {_CLASS = 'HashIndexTable'} do
-
---[[
-    The module 'bit' is available with:
-      * LuaJIT
-      * LuaBitOp extension which is available for:
-          * Lua 5.1
-          * Lua 5.2
-    The module 'bit32' is available with:
-      * Lua 5.2
-      * Lua 5.3 when compiled with LUA_COMPAT_5_2
-    The bitwise operators are added to Lua 5.3 as new lexemes (there causes
-    lexical error in older version)
---]]
-local band
-if _VERSION < 'Lua 5.3' then
-    local bit = bit32 or require'bit'
-    band = bit.band
-else
-    band = assert(load'--[[band]] return function (a, b) return a & b end')()
-end
-
-function HashIndexTable.new ()
-    local obj = {
-        hash_table = {length = 32;
-                      0, 0, 0, 0, 0, 0, 0, 0,
-                      0, 0, 0, 0, 0, 0, 0, 0,
-                      0, 0, 0, 0, 0, 0, 0, 0,
-                      0, 0, 0, 0, 0, 0, 0, 0},
-    }
-    return setmetatable(obj, {__index = HashIndexTable})
-end
-
-function HashIndexTable:add (name, index)
-    local slot = self:hash_slot_for(name)
-    if index < 255 then
-        -- increment by 1, 0 stands for empty
-        self.hash_table[slot] = band(index + 1, 0xFF)
-    else
-        self.hash_table[slot] = 0
-    end
-end
-
-function HashIndexTable:get (name)
-    local slot = self:hash_slot_for(name)
-    -- subtract 1, 0 stands for empty
-    return band(self.hash_table[slot], 0xFF) - 1
-end
-
-function HashIndexTable:string_hash (s)
-    -- this is not a proper hash, but sufficient for the benchmark,
-    -- and very portable!
-    return #s * 1402589
-end
-
-function HashIndexTable:hash_slot_for (element)
-    return band(self:string_hash(element), self.hash_table.length - 1) + 1
-end
-
-end -- class HashIndexTable
 
 function JsonObject.new ()
     local obj = {
